+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -7,6 +7,7 @@ export 'slack.dart';
 export 'telegram.dart';
 export 'tiktok.dart';
 export 'twitter.dart';
+export 'waze.dart';
 export 'whatsapp.dart';
 export 'youtube.dart';
 export 'zoom.dart';